requirements.txt
python-dotenv>=1.0.0
discord.py[voice]>=2.3.2
Flask>=3.0.0
pandas[all]
#numpy[all]